Crate governance_os_pallet_conviction_voting

This pallet can be used to deploy a conviction voting system. With conviction voting, votes accumulate and de-accumulate over time via a half life decay curve. It incentivizes people to vote earlier as their votes will have more value over time (up to a certain limit). In our implementation, we have decided that every new vote’s conviction will start at 0 and slowly accumulate to whatever the user staked on the proposal. This means that even if they stake a lot of coins they won’t be accounted immediately.

Re-exports

pub use $crate::traits::GetPalletVersion as _;

Structs

Conviction

Locks

Module

ProposalState

Proposals

VotingParameters

Enums

Call
Dispatchable calls.

Error

Constants

CONVICTION_VOTING_LOCK_ID

Traits

Config